requirements.txt
numpy
opencv-python
Pillow
ffmpeg-python
PyQt5
